requirements.txt
Django==4.0.6
django-allauth==0.50.0
django-cors-headers==3.10.1
django-crispy-forms==1.14.0
django-decorator-include==2.1
django-decouple==2.1
django-jsonfield-backport==1.0.4
django-sendgrid-v5==1.2.0
django-storages==1.12.3
django-tinymce==3.3.0
escapejson==1.0
gunicorn==20.1.0
Jinja2==3.0.3
#mysqlclient==2.1.0
Pillow==9.2.0
PyJWT==2.3.0
pyotp==2.7.0
pyparsing==3.0.8
python-decouple==3.6
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.2.1
tweepy==4.10.0
requests==2.25.1
ulid-py==1.1.0
urllib3==1.26.7
uvicorn==0.17.5
